+++ src/modules/verifiable-credentials/types/verifiable-credentials.types.ts
@@ -1,4 +1,5 @@
-import { IPresentationDefinition } from '@sphereon/pex';
+import { PresentationDefinition } from '@ew-did-registry/credentials-interface';
+import { IPresentationDefinition, SelectResults } from '@sphereon/pex';
 import { IssuerFields } from './role-credential.types';
 
 /*
@@ -7,21 +8,35 @@ import { IssuerFields } from './role-credential.types';
  */
 export interface RoleCredentialSubjectParams {
   /*
+   * DID of the subject
    * https://www.w3.org/TR/vc-data-model/#identifiers
    */
   id: string;
+
+  /* Role namespace */
   namespace: string;
+
+  /* Role version */
   version: string;
+
+  /* Role issuer fields */
   issuerFields?: IssuerFields[];
+
+  /* Expiration date of credential */
+  expirationDate?: Date;
 }
 
 export interface ProofOptions {
+  /* Proof verification method */
   verificationMethod?: string;
+
+  /* Proof purpose */
   proofPurpose?: string;
   challenge?: string;
 }
 
 export interface CreatePresentationParams {
+  /* Presentation definition from which pick the credentials */
   presentationDefinition?: IPresentationDefinition;
 }
 
@@ -29,4 +44,9 @@ export interface VerifyVerifiableCredentialResults {
   checks: string[];
   warnings: string[];
   errors: string[];
+}
+
+export interface InitiateExchangeResults {
+  presentationDefinition: PresentationDefinition;
+  selectResults: SelectResults;
 }